fixes bug#591994
10 Jun 2010 10:40:17 
daid@cns.s.u-tokyo.ac.jp

Bugs in KaliVeda

First bug is trivial.  In short, sanity checks deny a user from choosing a reaction which outputs a neutron.
There is no reason that a neutron is not a valid output of a reaction, and the patch just makes the sanity check use correct physics.
The physics appear to work fine with neutron reactions now.
This bug is in KVNucleus.cpp


The second bug is more serious.  The laboratory angle of the particles is not calculated with full relativity.
This is a problem in itself, but also results in internal collisions of the program, because the sanity checks are apparently using relativity, while the code is not.
In short, the angle calculation is missing a factor 1/gamma
This bug is in KV2Body.cpp

VERSION updated to 1.7.6b for immediate bugfix release