fixes bug#591994
10 Jun 2010 10:40:17
daid@cns.s.u-tokyo.ac.jp

The laboratory angle of the particles is not calculated with full relativity.
This is a problem in itself, but also results in internal collisions of the program, because the sanity checks are apparently using relativity, while the code is not.
In short, the angle calculation is missing a factor 1/gamma
This bug is in KV2Body.cpp