FIXED MASS FORMULA MESS:
KVNucleus objects are created with mass formula = beta stability
(this is no longer dependent on a .kvrootrc environment variable).

Default mass formula for KVINDRAReconNuc objects is EAL.

In KVReconstructedNucleus::SetIdentification, we only set the mass
of the nucleus if for the identification result both Z & A are measured.
Therefore for all nuclei with only Z measured, the mass will be
calculated from the mass formula of the nucleus (EAL for all
INDRA nuclei except E475S, for which EAL residue formula is used).