Post merge branch commit. Merged the old branch with KaliVeda 1.8.1.
N.B. I had to comment some lines in the following file to resolve
a conflict:

./KVIndra/calibration/KVLightEnergyCsI.cpp

I chose to keep the un-weighted 'lumcalc' variable as this is the one
present in the official KaliVeda version.  It can easily be switched
if the one weighted by A*Z*Z is needed.