* new KVNamedParameter is a general parameter class with a name, value can be integer, double or string; intended to replace the TNamed objects used in previous version of KVNameValueList, and also the KVParameter<Atype> template classes used in KVParameterList and KVGenParList.
* new version of KVNameValueList using KVNamedParameter to store values, no more
problems with floating-point precision, we store in a double

* corrected KVSeqCollection Streamer in order to be able to write and read these lists
(and the objects they contain) in a file without generating memory leaks.

* modified KVEvent::AddParticle to use (new?) method TClonesArray::ConstructedAt, in this way the ctors
of particles in an event are called only once during the lifetime of the array, when they are first allocated.

* added KVNameValueList to KVParticle and KVEvent to serve as general lists of parameters (int, double, string).
removed double inheritance of KVSimNucleus and KVSimEvent and rewrote to use the parameter lists which
are now in the base classes.

* modified the fGroups and fBoosted pointer members of KVParticle: they are now object members, as the use
of TClonesArray::ConstructedAt ensures that their ctor is called only once. This allows to remove lots of
costly 'if(!fGroups) fGroups = new KV...' stuff.

* tested writing new KVEvent/KVINDRAReconEvent objects in TTree on disk. There is no memory leak.
However, when performing event reconstruction on raw data, there is a leak corresponding to KVList/TList
objects (the following after reconstructing 500 events):

Warning in <KVClassMonitor::CompareToInit>: TList +928     (46672 --> 47600)
Warning in <KVClassMonitor::CompareToInit>: KVList +848     (23095 --> 23943)

The only KVList in the particle classes is the fBoosted member of KVParticle. Removing fBoosted from
the class structure does not change the leak. Therefore this must be due to a leak in the actual
event reconstruction.