INDRA_e503: KVHarpeeSi calibration file updated.
The old calibration file (Si_v3.cal) apparently did not include the run offset
(as in Sive503::Calibrate) so I've updated it and bumped the version number up
to 4. I'll leave the old file in place until I'm sure the new offsets are OK.
For all the runs I've looked at so far the timing now matches that of the old
recon code.