BackTrack: new classes GenericPDFModel and Simple1DModel  - models for backtracking based on analytic PDF's, and example    class for a very simple 1D model new example function simple1Dmodel_ex1.C  - demonstrates test of backtracking using Simple1DModel class
KVMultiDet:
new example (of KVRungeKutta and KVWilckeReactionParameters) 'ClassTraj'
 - simulate classical trajectories of heavy-ion collisions using
   Bondorf coulomb+nuclear potential, plus an ad-hoc friction term

Build system:
  Each subproject can have an 'examples' directory containing either
functions (in a single .C/.cpp/.cxx file) or classes (with a matching
.h file). Each example can use any of the classes in the subproject
or in any of the subprojects built before this one.
  Each example is built after all other modules of the subproject, this
is to help ensure that new changes do not break existing code. Only the
source for each example is installed, in one of:
  $(prefix)/examples                 - 'classic' install
  $(prefix)/share/kaliveda/examples  - 'gnuinstall'