fixes potential bug in KVLauncherGUI::AddButton
Class() is a static method: it returns the class of the pointer, not of
the object pointed to (if it is an object of a derived class). IsA()
gives the class of the object pointed to.

Test to see if required method is in object's public interface was
inverted & did exactly the opposite of what is written in comments.

The 'Connect' method would be called with the wrong class name in case
the obj argument is given.