KVTreeAnalyzer: enable PROOF for drawing histograms & making selections
PROOF can be used to speed up the generation of histograms.
For generation of selections, PROOF only works with the latest git versions
of ROOT 5.34 & 6.06, which include a bugfix to allow generation
of TEntryList with PROOF. With these versions, set the variable
   KVTreeAnalyzer.PROOFSelections:  on
in .kvrootrc in order to use it.
By default, this feature is disabled.