ROOT astyle config file & git commit hook to apply it
In order to automatically enforce coding conventions, we use a pre-commit hook
which will re-format all source code before it is committed using astyle
and the latest ROOT coding style. The hook will be set up by cmake if it is not
in place.

This means that if astyle is not installed, no source code can be committed
(other files are not restricted)
only prevent commits for lack of astyle if astyle should be used (source code)
automatically set up the git pre-commit hook on first configure/build