add method to know if an object was actually added to the list or not
when KVUniqueNameList::Add is called, we cannot assume that the list increases
in size: if it has the same name as another object in the list then either
it will not be added, or (if 'replace' mode is used) it will be added but
the previous object with the same name will be removed from the list
(and deleted if the list owns its objects). In both cases the size of the
list does not increase.