KVMultiDet: Minor post-merge fixes.
The KV2Body::GetIntegratedXSecRuthLab functions were uncommented in the
merge and so had to be commented again. A couple of the gui header files
were also missing a #include "TFunction.h". With these changes the
KVMultiDet part of the build is successful (using ROOT 5.34/32, I will
try ROOT 6 at some point).