VAMOS: Preliminary VAMOS layout (compiles).
The VAMOS directory as it exists at this point compiles OK but I'm not entirely
satisfied with the layout. As the new build system requires the code contained
within the subdirectories to compile into individual (modular) *.so libraries,
the dependency hierarchy must be treated properly (hence the folder labelled
'geom_deps' which is required to be built before 'geometry').