VAMOS: Rearranged VAMOS again.
I didn't like the idea of having a 'geom_deps' directory so I moved
those files back into 'geometry'. I also tweaked the build dependency
for db so that geometry builds first.

I realise that some of the source files for VAMOS are not where you
might expect them given the directory structure but this is simply
because the source code for VAMOS was not written with the build
dependencies as they are now (individual *.so libraries for each
directory). Now that we have migrated to the new build system this
situation should hopefully improve over time as the new build rules are
enforced.