Added an UNUSED() macro to hide unused function parameters.
To do this I've added a header file called KVMacros.h so that I can be
sure that the macro is defined identically everywhere in the code. It is
of course for general use and other global macros can be added if
needed.

Fixed several such situations in KVIndra/identification.