VAMOS: Bulk import of the E503 alterations.
This code is still being tested and at least one issue is known, there
is an over-estimation of A value occurring in the SiliconEnergyMinimiser
as a result of an incorrectly determined CsI energy (calibrator tends to
overshoot the energy for a given light input).

The old bisection algorithm has been removed in favour of a brute force
approach (See MonoMinimiserImpl.h). The SiliconEnergyMinimiser now acts
as a wrapper for the minimisation algorithm, allowing it to be altered
at whim. So long as all access to the minimiser implementations goes
through this class then any existing code should be unaffected by
changes in implementation.

As an example, there are currently two minimisation algorithms
available, one is the threaded version of the other. I recommend
sticking with the single threaded version (default) as it is less
complex and will no doubt be easier to debug. Within
SiliconEnergyMinimiser.h is a brief description of how to change the
implementation.

I believe all of the E503 changes (save a few bugs) have now been merged
with the main development tree. If I've left anything out I will be sure
to merge the appropriate changes as soon as I notice.

Note: Although I have tried to make the new classes compatible with
both C++11 and C++98, I have only tested compilation with C++11 and
ROOT6 at the moment.