VAMOS: Fixed the identification functions.
The KVVAMOSReconNuc::Identify() implementation has been removed from the
mainline KVVAMOSReconNuc class and so as a stop-gap measure I've
implemented a 'hack' function in Corrections_e503 to perform the
identification in the "old" 1.9 way. This is not intended to be a long
term solution.

Several fixes were also needed to the KVHarpeeCsI energy calculation
functions.