Add visualisation of tracks in multidetector geometry
You can now visualise (with the OpenGL viewer) tracks corresponding to simulated
detection of particles in the multidetector geometry.
Tracking is by default switched off: to enable and use it, do:

gMultiDetArray->GetNavigator()->SetTracking(kTRUE)
gMultiDetArray->DetectEvent(...)
gMultiDetArray->Draw("tracks")

The event will be drawn in the viewer overlaid on the geometry.