use list of KVRunFile objects for lists of available runs
change in KVAvailableRunsFile::GetListOfSystems(KVDBSystem* sys) when
sys!=0: this used to return a TList containing KVDBRun pointers, now it
returns a list of KVRunFile objects which belong to the list.
It is now possible to have several files associated with the same run
appear in KaliVedaGUI with their individual date, filename, version, etc.