temporary fix: give 'sps=1' option as default for all jobs
this is due to a problem with rootcling which automatically adds build
directory include paths to dictionaries. these paths have to be visible at
runtime. as KV is built on /sps/indra, the /sps disk must be available
for the batch job.

waiting for help from ROOT support to resolve this issue.