fixes (?) KVParticleCondition bug seen with ROOT6
when using several KVParticleCondition selections in an analysis with ROOT6,
the optimization can fail more or less randomly for one or more of the
conditions. as the condition is then systematically evaluated as false,
if the user does not read carefully all messages in the output of her
analysis, she may not realize there is a problem.

the problem seems to be solved by not deleting the generated .cpp and .h
files for the optimized class, but just in case, failure to optimize a
KVParticleCondition now produces a Fatal error, so it can not be ignored.