corrected (?) geometry for TESTLNS17 dataset...
...which is the geometry currently used for FAZIAPRE experiment,
FAZIAPRE dataset does not exist; however in the kaliveda-fazia.rootrc
file there are 3 lines describing a FAZIA geometry called FAZIAPRE
(with a minimum angle of 1 deg.), unfortunately (like the preceding lines
for FAZIASYM and FAZIACOR) these informations are apparently not used by
any part of the code, and I don't know what the point of them is...

Anyway, to cut a long story short, in the KVFAZIALNS17 geometry class
the minimum angle was set to 1.9 deg. and (compared to photo taken this
morning) the blocks look too spaced out (man). Putting the 'FAZIAPRE' value
of 1 deg. looks almost identical to the photo.

Maybe somebody should rationalize the way FAZIA geometries are handled?