kaliveda-sim: fix bug with ad hoc systems when filtering
when launching two successive jobs with different ad hoc systems, only the
first system was read in from the TGTextEntry, and the second job (and
any others) was performed using the same system (i.e. the same - wrong -
kinematics in case simulation was in CM frame).

now we update the fSystem variable every time something is typed in the
text box, or when a system is chosen in the drop-down list.
N.B. if a system is chosen in the list, and then an ad hoc system is
written in the text box, it is the latter which takes precedence.