correction to quality code assignment for KVIDZAFromZGrid
if isotope intervals are available for a Z but the data does not fall
inside one of the intervals, the quality code is set to IDCODE4 (bad).
this data is to be rejected.

in this way, for Z with isotopes a clean isotopic identification is
acheived, while for the other Z lines we identify with the same width
as a standard Z-only grid.